tween snowfall independently of map
The snowfall uses a custom update function to manually update each particle's x and y values.  The map uses Actor tweens.

The snowfall is no longer confined to only fall in the space that the screen's viewport shows but, rather, is constantly falling across the entire map.  I'm sure more clever + efficient solutions exist, but this implementation achieves the goal of having the snowfall not "follow" the map tweens.

It looks more natural now.  Thanks to Allen for noticing that the original implementation looked strange.